Make obstacle detection work with gazebo sim (realsense_runner.py WIP)
