+++ obstacle_detection/src/global_occupancy_grid.py
@@ -47,7 +47,6 @@ class GlobalOccupancyGrid:
         self.global_grid = np.zeros(self.global_grid_shape)
         self.global_counts = np.zeros_like(self.global_grid)  # keeps track of number of measurements for each cell
         self.global_totals = np.zeros_like(self.global_grid)  # keeps track of sum of measurements for each cell
-        self.localization_topic = rospy.get_param('localization_name')
         self.local_grid_topic = 'local_occupancy_grid'
         self.viz_dir = 'global_map/'
         self.data_dir = 'occupancy_grid_data/'
@@ -210,13 +209,15 @@ class GlobalOccupancyGrid:
 
                 fig.savefig('%s/%d' % (self.viz_dir, len(os.listdir(self.viz_dir))))
                 plt.close()
+        else:
+            print("No localization data, can not form global grid")
 
 
 if __name__ == '__main__':
     try:
         global_grid = GlobalOccupancyGrid()
         rospy.Subscriber(global_grid.local_grid_topic, OccupancyGrid, global_grid.local_grid_callback, queue_size=1)
-        rospy.Subscriber(global_grid.localization_topic, Odometry, global_grid.localization_listener, queue_size=1)
+        rospy.Subscriber("/odometry/filtered_map", Odometry, global_grid.localization_listener, queue_size=1)
         rospy.spin()
     except rospy.exceptions.ROSInterruptException:
         pass
